move UndoableCommandHistory
